+++ game_patch/multi/server_internal.h
@@ -38,6 +38,11 @@ struct HitSoundsConfig
     int rate_limit = 10;
 };
 
+struct OvertimeConfig
+{
+    bool enabled = false;
+    int additional_time = 5;
+};
 struct NewSpawnLogicConfig // defaults match stock game
 {
     bool respect_team_spawns = true;    
@@ -64,6 +69,7 @@ struct ServerAdditionalConfig
     std::optional<float> spawn_armor;
     int ctf_flag_return_time_ms = 25000;
     HitSoundsConfig hit_sounds;
+    OvertimeConfig overtime;
     std::map<std::string, std::string> item_replacements;
     std::string default_player_weapon;
     std::optional<int> default_player_weapon_ammo;