+++ docs/CHANGELOG.md
@@ -63,6 +63,7 @@ Version 1.9.0 (not released yet)
 - Add `server_password` command
 - Add `server_rcon_password` command
 - Make player maintain relative velocity when teleporting
+- Add `$DF Overtime Enabled` option in dedicated server config
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------